Refactored the merge
