+++ app.rb
@@ -29,7 +29,7 @@ class MakersBnb < Sinatra::Base
   end
 
   post '/spaces/new' do
-    @new_room = Room.new(new_space: params[:new_space]) 
+    @new_space = Room.new(new_space: params[:new_space]) 
   end
 
   get '/booking' do
@@ -43,7 +43,7 @@ class MakersBnb < Sinatra::Base
   end
   
   get '/list-spaces' do
-    erb :list_space
+    erb :new
   end
 
   run! if app_file == $0
