--- lib/spaces.rb
@@ -1,8 +1,0 @@
-class Spaces
-  def self.all
-    ['Haunted hill',
-    'Trump tower',
-    'The Corleones',
-    'Bates Mansion']
-  end
-end
