+++ spec/features/list_space_spec.rb
@@ -1,6 +1,6 @@
 feature 'list space' do
   scenario 'Viewing a list of spaces' do
     visit '/list-spaces'
-    expect(page).to have_content 'List your space!'
+    expect(page).to have_content 'List a space'
   end
 end
